[BOOK-150] refactor: :screens 모듈 feature 모듈 내부로 이동
번복
