+++ feature/library/src/main/kotlin/com/ninecraft/booket/feature/library/LibraryScreen.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
 import com.ninecraft.booket.core.designsystem.DevicePreview
 import com.ninecraft.booket.core.designsystem.theme.ReedTheme
-import com.ninecraft.booket.screens.LibraryScreen
+import com.ninecraft.booket.feature.screens.LibraryScreen
 import com.slack.circuit.codegen.annotations.CircuitInject
 import dagger.hilt.android.components.ActivityRetainedComponent
