+++ feature/search/src/main/kotlin/com/ninecraft/booket/feature/search/SearchPresenter.kt
@@ -10,7 +10,7 @@ import androidx.compose.runtime.setValue
 import com.ninecraft.booket.core.data.api.repository.BookRepository
 import com.ninecraft.booket.core.model.BookSearchModel
 import com.ninecraft.booket.core.model.BookSummaryModel
-import com.ninecraft.booket.screens.SearchScreen
+import com.ninecraft.booket.feature.screens.SearchScreen
 import com.orhanobut.logger.Logger
 import com.slack.circuit.codegen.annotations.CircuitInject
 import com.slack.circuit.retained.rememberRetained
