+++ settings.gradle.kts
@@ -38,8 +38,7 @@ include(
     ":feature:library",
     ":feature:login",
     ":feature:main",
+    ":feature:screens",
     ":feature:search",
     ":feature:settings",
-
-    ":screens",
 )
